requirements.txt
pyyaml
logging
setuptools
wheel
twine
